requirements.txt
streamlit
python-dotenv
openai
pdfplumber
pandas
